Fixed incorrect reporting of beforeEach, and double time out issue
beforeEach and afterEach functions were marked "triggered" or "timedOut" on the function object itself, which caused the flags to be ovewritten every time they ran. Now they report to spec.results.befores and ..afters directly.

A double timeout issue was caused by assigning the timeout to the context of serial execution of befores-spec-afters clause - which should never have a timeout as it is a fully internal process.